Consolidate pacakge checks for now
