+++ src/packages/shared-utils/package-actions/rules.ts
@@ -1,10 +1,6 @@
-import {
-  Action,
-  ActionAvailabilityCheck,
-  ActionRule,
-  SEATOOL_STATUS,
-} from "../../shared-types";
+import { Action, ActionRule, SEATOOL_STATUS } from "../../shared-types";
 import { isCmsUser, isStateUser } from "../user-helper";
+import { PackageCheck } from "../packageCheck";
 
 const arIssueRai: ActionRule = {
   action: Action.ISSUE_RAI,
